mt8163:fix build error and apply more changes for mt8163 Signed-off-by: Kai Jones <joneskai626@gmail.com> Change-Id: I42a9c8bae2ac4e8259149de8408c86a303dc6eb5

+++ media/libmedia/IOMX.cpp
@@ -672,7 +672,7 @@ status_t BnOMXNode::onTransact(
 
             OMX_INDEXTYPE index = static_cast<OMX_INDEXTYPE>(data.readInt32());
 //test
-            size_t size = data.readInt32();
+            size_t size = data.readInt64();
 
             status_t err = NOT_ENOUGH_DATA;
             void *params = NULL;
